requirements.txt
django==3.2
python-dotenv==0.19.2
django_split_settings==1.1.0
django-debug-toolbar==3.2.4
psycopg2-binary==2.9.3
marshmallow==3.14.1
dotenv-python==0.0.1
gunicorn==20.1.0